+++ src/components/article/SubjectListing/subjectListing.scss
@@ -1,6 +1,7 @@
 @import "src/lib.scss";
 
 .subject-listing {
+  margin-top: 3rem;
   &__section {
     margin-top: 3rem;
     &:first-child {
