--- src/pages/universalPage/universalPage.basic.stories.js
@@ -1,13 +1,0 @@
-import universalPageBasic from './universalPageBasic.html.twig';
-
-export default {
-  title: 'Pages/Universal Page/Basic',
-  parameters: {
-    controls: { disabled: true },
-    layout: 'fullscreen',
-  },
-};
-
-export const Basic = () => {
-  return universalPageBasic();
-};
