+++ src/pages/universalPage/universalPage.stories.js
@@ -1,0 +1,13 @@
+import universalPage from './universalPage.html.twig';
+
+export default {
+  title: 'Pages/Universal Page',
+  parameters: {
+    controls: { disabled: true },
+    layout: 'fullscreen',
+  },
+};
+
+export const UniversalPage = () => {
+  return universalPage();
+};
